+++ lib/gpos720_printer_platform_interface.dart
@@ -57,6 +57,13 @@ abstract class Gpos720PrinterPlatform extends PlatformInterface {
         'imprimirImagem(Uint8List data, int width, int height, AlignmentTypes align) has not been implemented.');
   }
 
+  Future<PrinterStatus> imprimirImagemFiltrada(
+      Uint8List data, int width, int height,
+      {AlignmentTypes align = AlignmentTypes.center, double? threshold}) {
+    throw UnimplementedError(
+        'imprimirImagemFiltrada(Uint8List data, int width, int height, AlignmentTypes align, double? threshold) has not been implemented.');
+  }
+
   Future<PrinterStatus> imprimirCodigoDeBarra(
       String mensagem, int width, int height, BarcodeTypes barcodeType) {
     throw UnimplementedError(